[chore] kafkareceiver: fix flaky test (#43385)
#### Description

Ensure the receiver consumes both records in a single fetch, since the
test is verifying behaviour of fetch processing during shutdown. If the
records are split across fetches, then the second record will not be
processed since the second fetch will fail due to the client having been
shut down.

Simplify the test while we're at it: there's no need to wait for
partition assignment, or checking the number of calls to the consumer;
sending to the "consuming" channel is good enough.

#### Link to tracking issue

Fixes #43116

#### Testing

Run the test in a loop. Also, see
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/43116#issuecomment-3384081516

#### Documentation

N/A