[chore] Move Vihas Makwana to approvers (#43439)
Vihas is an active triager for the last 11 months. This PR offers to
elevate him to the status of approver.

Vihas has been a consistent active member of this community. Below is a
snapshot of his activity:

Issues interacted with:
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues?q=is%3Aissue%20state%3Aopen%20commenter%3AVihasMakwana
PRs reviewed:
https://github.com/open-telemetry/opentelemetry-collector-contrib/pulls?q=is%3Apr+reviewed-by%3AVihasMakwana+

It is my pleasure to open this proposal, and I want to thank Vihas for
his contributions so far, and I look forward to his impact in this new
role.

@VihasMakwana please confirm that you would like to be considered for
this role.

@open-telemetry/collector-contrib-approvers please review and approve in
support.