[chore] Fix tools for windows 11 arm (#43633)
Fix #43625

[Updated to correctly describe the problem]

The tools used in the `lint` target need to build the component and per
manual tests the component fails to build if `go build` tries to use
`CGO`

The issue didn't show up previously because the build and test workflow
on Windows ARM already sets the environment variable `CGO_ENABLED` to
`0`. The fix here is to adopt the same approach and disable `CGO` for
`windows-11-arm`.

For safety I will do a bogus change to trigger a run of the case that
exposed the issue.