[chore][ci-cd] Fix `mod-integration-test` target (#43702)
Thanks to @jkoronaAtCisco for calling my attention to this issue.

The expression building the name of the tests has an extra white-space
and this caused the integration tests to not run the first integration
test captured in the expression. For packages with a single integration
test this means that they didn't run at all since the issue was
introduced. It is possible that with this fix some integration test
surface.

The issue is just an extra space in the make file that should be
removed.