Move metricstarttimeprocessor to beta (#43912)
#### Description

Move the metricstarttime processor to beta. We have been using it in
production, and have fixed issues that have come up. At this point, we
don't see needing any breaking chagnes, so it is time to move it to
beta, and replace the adjustment in the prometheus receiver.

#### Link to tracking issue
Part of
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/43656

@ridwanmsharif