[receiver/prometheusreceiver] fix stalness tracking (#43925)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

In the dependency upgrade #43890 there was the PR
https://github.com/prometheus/prometheus/pull/16429 which changed the
provisions for staleness tracking. Now the code only does the tracking
if the series was successfully appended in storage. This is indicated by
a non zero storage reference returned by the appender.

Since we used to return 0 in all cases, the staleness tracking is now
broken.

Solution is to return a fake reference of 1, just to indicate success.

Alternative solution would be to make a unique reference for each series
label set, but that incurs an overhead which we can avoid.

Performance impact: probably small as the cache should be stable over
time - unless targets churn a lot between receiver instances.

Resource utilization: +1 hashmap with series identifier (name+labels)
pointing to cache entries. Size related to number of series. `addRef`
here:
https://github.com/prometheus/prometheus/blob/c8f1de18a7870e5dc99ed50119c194691c41e41c/scrape/scrape.go#L1089

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #43893

<!--Describe what testing was performed and which tests were added.-->
#### Testing

Un-skip related tests.

---------

Signed-off-by: György Krajcsovits <gyorgy.krajcsovits@grafana.com>