[chore] Fix First Time Contributor flow (#44205)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
It's very nice of the bot to welcome first time contributors, but it's a
bit too overzealous about it. :)

The current state of the action causes it to run every time any labels
on the target PR are changed. This PR updates the action it so it only
runs on PR open.


<!--Describe what testing was performed and which tests were added.-->
#### Testing
I'm actually unsure how to test this, but I am reasonably sure this will
be fine (famous last words I guess). It appears to always work with the
PR opening, so I don't think the labelling is necessary. It looks like
the `label` trigger was added to support manually adding the label, but
it doesn't appear that will be necessary since the PR open trigger
handles it.


<!--Please delete paragraphs that you did not use before submitting.-->