[receiver/sqlserver] Fix incorrect usage of topQueryCount instead of MaxQuerySampleCount (#44304)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
In the `sqlserverreceiver`, for the query that fetches query metrics and
plans, the parameters are passed from the receiver. In this code,
`TopQueryCollection.topQueryCount` was used instead of
`TopQueryCollection.MaxQuerySampleCount`, causing a significantly
smaller number of queries to be sampled for calculating the actual top N
results.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/44303

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Unit tests updated