[chore][pkg/ottl] improve linting (#44258)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

This is part of the process to enable the
[modernize](https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/44199)
linter in CI and upgrade `golanglint` to the latest version.

There are no changes in the component behaviour.

```sh
pkg/ottl/expression_test.go:974:5: error-is-as: use assert.ErrorIs or assert.ErrorAs depending on the case (testifylint)
				assert.IsType(t, TypeError(""), err)
				^
pkg/ottl/expression_test.go:1224:5: error-is-as: use assert.ErrorIs or assert.ErrorAs depending on the case (testifylint)
				assert.IsType(t, TypeError(""), err)
				^
pkg/ottl/expression_test.go:1301:5: error-is-as: use assert.ErrorIs or assert.ErrorAs depending on the case (testifylint)
				assert.IsType(t, TypeError(""), err)
				^
pkg/ottl/expression_test.go:1478:5: error-is-as: use assert.ErrorIs or assert.ErrorAs depending on the case (testifylint)
				assert.IsType(t, TypeError(""), err)
				^
pkg/ottl/expression_test.go:1555:5: error-is-as: use assert.ErrorIs or assert.ErrorAs depending on the case (testifylint)
				assert.IsType(t, TypeError(""), err)
				^
pkg/ottl/expression_test.go:1732:5: error-is-as: use assert.ErrorIs or assert.ErrorAs depending on the case (testifylint)
				assert.IsType(t, TypeError(""), err)
				^
pkg/ottl/expression_test.go:1931:5: error-is-as: use assert.ErrorIs or assert.ErrorAs depending on the case (testifylint)
				assert.IsType(t, TypeError(""), err)
				^
pkg/ottl/expression_test.go:2008:5: error-is-as: use assert.ErrorIs or assert.ErrorAs depending on the case (testifylint)
				assert.IsType(t, TypeError(""), err)
				^
pkg/ottl/expression_test.go:2164:5: error-is-as: use assert.ErrorIs or assert.ErrorAs depending on the case (testifylint)
				assert.IsType(t, TypeError(""), err)
				^
pkg/ottl/expression_test.go:2361:5: error-is-as: use assert.ErrorIs or assert.ErrorAs depending on the case (testifylint)
				assert.IsType(t, TypeError(""), err)
				^
pkg/ottl/expression_test.go:2447:5: error-is-as: use assert.ErrorIs or assert.ErrorAs depending on the case (testifylint)
				assert.IsType(t, TypeError(""), err)
				^
pkg/ottl/ottlfuncs/func_is_list_test.go:189:2: error-is-as: use assert.ErrorIs or assert.ErrorAs depending on the case (testifylint)
	assert.IsType(t, ottl.TypeError(""), err)
	^
make[2]: *** [lint] Error 1
make[1]: *** [pkg/ottl] Error 2
```

Signed-off-by: Paulo Dias <paulodias.gm@gmail.com>