prepare release v0.140.1 (#44354)
Replaces
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/44346.

After release `v0.140.0` was tagged in Contrib, the [integration tests
failed](https://github.com/open-telemetry/opentelemetry-collector-contrib/actions/runs/19438560491),
causing the release to not be published. Since the tags are immutable, I
prepared a new release `v0.140.1` of Contrib, which includes the fix to
integration tests
https://github.com/open-telemetry/opentelemetry-collector-contrib/commit/6d5f384830951858b616bb4ff5d787ce2a3bb381.
After that, I manually ran some pre-release commands based on the
[prepare-release](https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/v0.140.0/.github/workflows/prepare-release.yml#L19)
workflow and the
[release-prepare-release.sh](https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/v0.140.0/.github/workflows/scripts/release-prepare-release.sh)
script used in it. See individual commits for the exact commands.

---------

Co-authored-by: Yang Song <songy23@users.noreply.github.com>