Add exponential histogram support to the cumulative-to-delta processor (#44106)
#### Description

This PR adds support for exponential histograms to the
cumulative-to-delta processor, which currently lacks it.

Along the way, I also:
- made a small change in `pdatatest` so that `nil` and an empty slice
are considered equal for exponential histogram buckets, since (as far as
I know) they have the same semantics. This is somewhat necessary because
a newly-created `pcommon.UInt64Slice` will be nil, but once filled,
there is no way to reset it back to nil: the best you get is an empty
slice.
- modified `convertHistogramDataPoints` to take a typed argument instead
of `any`, like was done in #35166.

#### Link to tracking issue
N/A

#### Testing
I added two simple test cases to the existing test.

#### Documentation
I updated the README to reflect the added support.