[receiver/azureeventhub] Promote UseAzeventhubs feature gate to Beta (#44340)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

This PR promotes the feature gate
`receiver.azureeventhubreceiver.UseAzeventhubs` to Beta as no issues
seem been reported for the time this feature was in Alpha stage.

This is also a preparation for future removal usage of the Go module
[github.com/Azure/go-autorest](https://github.com/Azure/go-autorest)
since it has been deprecated.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Relates to #44335

<!--Describe the documentation added.-->
#### Documentation

Updated documentation accordingly

<!--Please delete paragraphs that you did not use before submitting.-->

---------

Signed-off-by: Paulo Dias <paulodias.gm@gmail.com>