[chore][translator/azurelogs] Regenerate `*_expected.yaml` testing files (#44267)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

This PR is to regenerate the `*_expected.yaml` testing files using
`golden.WriteLogs(t, expectedFile, logs)` in
`resourcelogs_to_logs_test.go`.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue

There is no issue for this. I'm putting it in a separate PR so it can be
more clear
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/44266
does not break the other existing log categories.