feat: Add macOS Unified Logging Receiver - Config/Factory (#44100)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
Adds the factory and configuration implementation for the macOS Unified
Logging receiver. This PR follows the steps outlined in
`CONTRIBUTING.md` to split up the PRs for a new component.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
#44089 

<!--Describe what testing was performed and which tests were added.-->
#### Testing
Unit tests for the configuration and factory.

<!--Describe the documentation added.-->
#### Documentation
The receiver is completely documented in `README.md`.

<!--Please delete paragraphs that you did not use before submitting.-->