fix(ottl): enforce valid types for setters in contexts (#43505)
#### Description
Introduce a common utility to check for the expected type and returns a
consistent error message.
Also applies the use of this utility to ctxscope, ctxlog and ctxmetric.

#### Link to tracking issue
Fixes #40198

#### Testing
Adds testing for the new utility, will address any tests whose behavior
is changed by this propagation of error that didn't exist before.

#### TODO
As it is now, the PR changes some but not all setters, I can either
apply them all here or create a followup PR with more refactorings.

#### Alternate designs
Introducing a new helper in ctxerror was considered but it can be done
at a later stage, as it requires consensus on the set of metadata that
such helper would require or optional (e.g. context name, path segment).