[pkg/ottl] Add `ParseSeverity` function (#37280)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

This PR adds the `ParseSeverity` function, as discussed in the linked
ticket. I also had to make a minor change to the
internal `mapGetter`, handling the map literals to return a raw
`map[string]any`, instead of a `pcommon.Map`. This is because if there
is a map literal within a slice, the `pcommon.Slice.FromRaw` cannot
handle the `pcommon.Map`, as it only works with raw data types.

This change is however transparent, and the behavior to the outside of
this package does not change.
EDIT: After merging main with the support for value expressions,
introduced in #36883, this would affect the type of values returned by
`ParseValueExpression` - previously this could potentially return
`map[string]any`/`[]any`, but with the changes introduced in this PR,
this would return a `pcommon.Map`/`pcommon.Slice`.
Please let me know  if I should do this change in a separate PR though.



<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #35079

<!--Describe what testing was performed and which tests were added.-->
#### Testing

Added unit and e2e tests

<!--Describe the documentation added.-->
#### Documentation

Describe new function in the readme

<!--Please delete paragraphs that you did not use before submitting.-->

---------

Signed-off-by: Florian Bacher <florian.bacher@dynatrace.com>
Co-authored-by: Tyler Helmuth <12352919+TylerHelmuth@users.noreply.github.com>