[chore][receiver/googlecloudpubsub] update to cloud.google.com/go/pub… (#44466)
#### Description
Update to the cloud.google.com/go/pubsub/v2 package to keep the Renovate
bot happy.

This PR is scoped to the receiver/googlepubsubreceiver

#### Link to tracking issue
This PR fixes renovate trying to upgrade but can't because it requires
code changes:

https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/42871

And remove the Pubsub part in the combined deprecation fix PR:

https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/41646

#### Testing
Changed existing test to match the new API.

#### Documentation
No documentation change required.