[chore] [receiver/k8sobjects] Extract logic for watching/pulling K8s resources into separate package (#43108)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

This PR refactors the `k8sobjects` receiver by extracting the components
responsible for receiving the K8s resources in the cluster into a
separate package with an `Observer` interface, and the pull/watch based
implementations thereof.
This package can then also be used by the `k8sevents` receiver, which
would then only need to format the received events to conform with the
semantic conventions.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Part of #40825 - next steps as part of this issue would be to also make
use of the observer components in the k8seventsreceiver

<!--Describe what testing was performed and which tests were added.-->
#### Testing

Added tests for the new components. The tests for the receiver itself
could be refactored to not use the concrete implementations of the
Observer interface, but that should probably be done as a follow up to
not make this PR too large

---------

Signed-off-by: Florian Bacher <florian.bacher@dynatrace.com>