[receiver/azuremonitorreceiver] Collect only supported aggregations for each metric (501 not implemented issue) (#43880)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

When we collect metrics, we ask for all available aggregations (Average,
Sum, Min, Max, ...) but when we do so with some metrics, Azure API is
giving a 501 Not Implemented error. So we need to really respect the
definition of each metric.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Relates #43648 (not fully fixing)

<!--Describe what testing was performed and which tests were added.-->
#### Testing
For the full e2e tests, the impact on tests has only been to setup
supported aggregation in the ``metrics definition`` mocks. We simulate
supporting all aggregation in the definition to make sure the tests
results are unchanged.

We rely on unit test of the ``getMetricAggregations`` function to test a
more variety of cases.

<!--Describe the documentation added.-->
#### Documentation

Updated to provide more details.
<img width="1089" height="829" alt="image"
src="https://github.com/user-attachments/assets/28b8559b-0b7a-4e23-9072-f4023ea61747"
/>

<!--Please delete paragraphs that you did not use before submitting.-->

Signed-off-by: Célian Garcia <celian.garcia@amadeus.com>
Signed-off-by: Celian GARCIA <celian.garcia@amadeus.com>