[chore] Ensure "update-otel workflow failed" issues have logs from all job steps (#44583)
#### Description

When the `update-otel` workflow fails, an issue titled "update-otel
workflow failed" is created, containing the last 100 lines of the CI's
log output to help diagnose the problem. However, we've recently seen
failures in steps that aren't covered by this log capture, which makes
the output useless. (Example: #44580, where the failure was while
creating the PR)

This PR ensures we capture logs from all steps in the `update-otel` job.