[receiver/systemd] Promote systemd receiver to alpha (#44647)
#### Description
This promotes the systemd receiver to alpha stability. While this is by
no means feature complete (e.g. per-unit CPU and memory usage, unit
uptime), its current functionality is still useful (for observability
around unit failures), and it's clear there is demand for it (e.g.
#44420).

There are still some open questions about the shape of the exported
metrics, such as whether to represent each unit as a resource or not (as
discussed in #33532, and what units to default to (see #44646), but I
think we'll be better able to answer those once this is used in the
wild.

#### Documentation
Additional examples have been added to the receiver's README.