[chore][receiver/awslambdareceiver] improve linting (#44778)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

Improve linting for `receiver/awslambda`.

This is part of the process to enable the
[modernize](https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/44199)
linter in CI.

There are no changes in the component behaviour.

```sh
Error: receiver/awslambdareceiver/internal/s3_test.go:64:3: error-is-as: use require.ErrorIs or require.ErrorAs depending on the case (testifylint)
		require.IsType(t, &consumererror.Error{}, err)
		^
```

Signed-off-by: Paulo Dias <paulodias.gm@gmail.com>