[receiver/awss3] Add `s3_partition_format` and align config options with S3 Exporter (#44759)
#### Description
This is a breaking change as the `s3_partition` config option has been
removed in favour of the `s3_partition_format` option. Along with the
`s3_partition_timezone` this now more closely aligns the configuration
options with the S3 Exporter. Finally add the ability to exclude the
telemetry type from the S3 key, using the new option
`file_prefix_include_telemetry_type`.

The `s3_partition_format` defaults to the format used for the default of
the previous `s3_partition` meaning that only those using the `hour`
setting for `s3_partition` should be impacted.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes #43720

<!--Describe what testing was performed and which tests were added.-->
#### Testing
New unit tests added

<!--Describe the documentation added.-->
#### Documentation
Documentation updated with the new configuration options

<!--Please delete paragraphs that you did not use before submitting.-->