[processor/tailsampling] Fix memory leak (#44884)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description
When the TSP is configured to drop samples on overflow we add each trace
id to a queue to drop, but if we never actually overflow the trace ids
are not removed from the queue causing a memory leak. Whenever a trace
is removed from the map we also need to drop it from the queue.

This was introduced in
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/43671
which was released in v0.141.0 but that is all so far.