[receiver/dockerstats] Upgrade default Docker API version to 1.44 (#44931)
#### Description

Latest version of the docker engine requires 1.44 as the minimum API
version. The effect meaning that default configuration is incompatible
with latest docker versions.

The minimum supported version does not change.

Previously I thought this PR would update this, but it was missed:
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/44411

#### Link to tracking issue
Fixes #44279 

#### Testing

I added coverage for testing with both minimum and default API versions
(before the version was hardcoded). This is mainly to ensure the docker
client operates the same in both scenarios.


#### Documentation

Updated README and made some slight corrections to minimum supported
version (it incorreectly said 1.22, where it is actually 1.25)