[connector/count] Fix regression in basic config behavior (#44540)
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
#### Description

See
[issue](https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/41769)
for details.

*TLDR:* `count` relies on
[`Unmarshal`](https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/4878a1c37b31b8610e2820398ec1150b565d556c/connector/countconnector/config.go#L150-L177)
to set default subconfigs, but [this
update](https://github.com/open-telemetry/opentelemetry-collector/commit/9ab25f6f66ae76aaef69acca4752dee2536b4e65)
changed confmap behavior to NOT call unmarshal for nil subconfigs like
this one which `count` expects to work:
```
connector:
  count:
```

so no metrics are emitted, even though this is a basic example of usage
from the
[readme](https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/v0.128.0/connector/countconnector/README.md#default-configuration).

A workaround is to set this instead:
```
connector:
  count: {}
```

but I think it's better to match the previous behavior by setting the
default configs with `createDefaultConfig` and just overriding them in
`Unmarshal` if custom metrics are defined.

<!-- Issue number (e.g. #1234) or full URL to issue, if applicable. -->
#### Link to tracking issue
Fixes
[41769](https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/41769)

<!--Describe what testing was performed and which tests were added.-->
#### Testing
- Added unit tests. Tests relevant to this PR's changes fail if run on
main. (Specifically tests 1-2)
- Manually tested the same configs covered by unit tests.
- There is a detailed test example in the issue.

<!--Describe the documentation added.-->
#### Documentation

N/A, this fix makes the behavior match documentation again.

<!--Please delete paragraphs that you did not use before submitting.-->